add sin, logn, log2, log3, log10

+++ src/main/java/ru/ifmo/tpo/lab2/funcions/IterationFunc.java
@@ -1,14 +1,13 @@
 package ru.ifmo.tpo.lab2.funcions;
 
-import java.math.BigDecimal;
 
-import ru.ifmo.tpo.lab2.intefaces.IFunc;
+import ru.ifmo.tpo.lab2.models.IFunc;
 
 public abstract class IterationFunc implements IFunc{
     private static final int MAX_ITERATIONS = 1000;
     
     @Override
-    public BigDecimal solve(BigDecimal x, BigDecimal accuracy) {
-        return BigDecimal.valueOf(MAX_ITERATIONS);
+    public Double solve(Double x, Double accuracy) {
+        return Double.valueOf(MAX_ITERATIONS);
     }
 }
